+++ server/queue/__init__.py
@@ -1,0 +1,1 @@
+"""Request queue package."""